Now it works! Numerous bug fixes, few new assertions, got rid of null context.

+++ com.yoursway.sadr.python.core/src/com/yoursway/sadr/python_v2/model/builtins/LongType.java
@@ -60,6 +60,6 @@ public class LongType extends PythonClassType {
     
     @Override
     public String describe() {
-        return "int";
+        return "long";
     }
 }
